requirements.txt
django==5.2
requests==2.32.3
djangorestframework-3.16.0